description_ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
google-genai>=0.2.0
boto3==1.34.28
Pillow==10.2.0
PyMySQL==1.1.0

# Legacy (Claude API, 필요시)
# anthropic==0.18.1
